Highlight detected matches for text exercises
