--- src/main/webapp/app/exercises/shared/plagiarism/plagiarism-split-view/programming-submission-viewer/programming-submission-viewer.component.html
@@ -1,11 +1,0 @@
-<jhi-split-pane-header [files]="files" (selectFile)="handleFileSelect($event)" studentLogin="{{ plagiarismSubmission?.studentLogin }}"></jhi-split-pane-header>
-
-<div class="programming-submission-viewer" [innerHTML]="fileContent || ' '" [ngClass]="{ 'no-file': !fileContent }" *ngIf="!loading">
-    {{ 'artemisApp.plagiarism.no-file-selected' | translate }}
-</div>
-
-<div class="plagiarism-submission-loader" *ngIf="loading">
-    <div class="spinner-border text-primary" role="status">
-        <span class="sr-only">Loading...</span>
-    </div>
-</div>
