--- src/main/webapp/app/exercises/shared/plagiarism/plagiarism-split-view/programming-submission-viewer/programming-submission-viewer.component.scss
@@ -1,23 +1,0 @@
-.programming-submission-viewer {
-    height: 100%;
-    overflow-y: auto;
-    padding: 0.5rem;
-    font-family: monospace;
-    white-space: pre;
-
-    &.no-file {
-        opacity: 0.5;
-    }
-}
-
-.plagiarism-match {
-    color: dodgerblue;
-    font-weight: bold;
-}
-
-.plagiarism-submission-loader {
-    align-items: center;
-    display: flex;
-    height: 100%;
-    justify-content: center;
-}
