+++ src/main/webapp/app/exercises/shared/plagiarism/plagiarism-split-view/text-submission-viewer/text-submission-viewer.component.html
@@ -1,9 +1,5 @@
-<jhi-split-pane-header studentLogin="{{ plagiarismSubmission?.studentLogin }}"></jhi-split-pane-header>
+<jhi-split-pane-header [files]="files" (selectFile)="handleFileSelect($event)" studentLogin="{{ plagiarismSubmission?.studentLogin }}"></jhi-split-pane-header>
 
-<div class="text-submission-viewer" *ngIf="!loading">{{ this.submission?.text }}</div>
-
-<div class="plagiarism-submission-loader" *ngIf="loading">
-    <div class="spinner-border text-primary" role="status">
-        <span class="sr-only">Loading...</span>
-    </div>
+<div class="text-submission-viewer" [innerHTML]="fileContent || ' '" [ngClass]="{ 'no-file': !fileContent, 'is-code': isProgrammingExercise }" *ngIf="!loading">
+    {{ 'artemisApp.plagiarism.no-file-selected' | translate }}
 </div>
