+++ src/main/webapp/app/exercises/shared/plagiarism/plagiarism-split-view/text-submission-viewer/text-submission-viewer.component.scss
@@ -2,7 +2,20 @@
     height: 100%;
     overflow-y: auto;
     padding: 0.5rem;
-    white-space: pre;
+    white-space: pre-wrap;
+
+    &.no-file {
+        opacity: 0.5;
+    }
+
+    &.is-code {
+        font-family: monospace;
+    }
+}
+
+.plagiarism-match {
+    color: dodgerblue;
+    font-weight: bold;
 }
 
 .plagiarism-submission-loader {
